(cleanup): Refactored some of the code to transition away from class components to functional components (hooks)
