+++ src/pages/Admin/Create/components/UploadUser/Panel/ManageUsersTab.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useReducer, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { Query, Mutation } from 'react-apollo'
 import gql from 'graphql-tag'
@@ -98,210 +98,208 @@ TeamSelect.propTypes = {
   handleChange: PropTypes.func.isRequired,
 }
 
-export default class ManageUsersTab extends React.Component {
-  state = { teamId: null, username: null, email: null, checkedItems: new Map(), transferId: null }
-
-  handleSelect = e => {
-    this.setState({ teamId: e.target.value })
-  }
-
-  handleInputChange = e => {
-    this.setState({
-      [e.target.name]: e.target.value,
-    })
-  }
-
-  handleCheckBox = e => {
-    const { name, checked } = e.target
-    this.setState(prevState => ({
-      checkedItems: prevState.checkedItems.set(name, checked),
-    }))
-  }
-
-  handleTransferSelect = e => {
-    this.setState({
-      transferId: e.target.value,
-    })
-  }
-
-  render() {
-    const { username, email, teamId, transferId, checkedItems } = this.state
-    const { eventId } = this.props
-
-    return (
-      <div>
-        <TeamSelect eventId={eventId} teamId={teamId} handleChange={this.handleSelect} />
-        <Query query={USER_LIST} variables={{ teamId }}>
-          {({ data, loading, error }) => {
-            if (!data) return null
-            if (loading) return <div>Loading...</div>
-            if (error) return <div>Error...</div>
-
-            if (teamId === null) {
-              return <H4>Select a user from the dropdown</H4>
-            }
-
-            if (!Array.isArray(data.user_team) || !data.user_team.length) {
-              return (
-                <table style={{ width: '100%', marginTop: 10 }}>
-                  <thead>
-                    <td>Username</td>
-                    <td>Email</td>
-                    <td />
-                  </thead>
-                  <tbody>
-                    <tr>
-                      <td>
-                        <InputGroup
-                          name="username"
-                          value={username}
-                          onChange={this.handleInputChange}
-                        />
-                      </td>
-                      <td>
-                        <InputGroup name="email" value={email} onChange={this.handleInputChange} />
-                      </td>
-                      <td>
-                        <Mutation
-                          mutation={ADD_USER_TO_TEAM}
-                          refetchQueries={[{ query: USER_LIST, variables: { teamId } }]}
-                        >
-                          {insert_user_team => (
-                            <Button
-                              intent="primary"
-                              icon={IconNames.PLUS}
-                              fill
-                              onClick={() => {
-                                insert_user_team({
-                                  variables: {
-                                    email,
-                                    username,
-                                    teamId,
-                                  },
-                                })
-                                this.setState({
-                                  email: '',
-                                  username: '',
-                                })
-                              }}
-                            />
-                          )}
-                        </Mutation>
-                      </td>
-                    </tr>
-                  </tbody>
-                </table>
-              )
-            }
-
-            return (
-              <div style={{ marginTop: 10 }}>
-                <table style={{ width: '100%' }}>
-                  <thead>
-                    <td>Username</td>
-                    <td>Email</td>
-                    <td />
-                  </thead>
-                  <tbody>
-                    {data.user_team.map(({ user }) => (
-                      <tr key={user.uuid}>
-                        <td>{user.username}</td>
-                        <td>{user.email}</td>
-                        <td>
-                          <Checkbox
-                            name={user.uuid}
-                            checked={checkedItems.get(user.uuid)}
-                            onChange={this.handleCheckBox}
-                            large
-                            style={{ width: '100%' }}
-                          />
-                        </td>
-                      </tr>
-                    ))}
-                    <tr>
-                      <td>
-                        <InputGroup
-                          name="username"
-                          value={username}
-                          onChange={this.handleInputChange}
-                        />
-                      </td>
-                      <td>
-                        <InputGroup name="email" value={email} onChange={this.handleInputChange} />
-                      </td>
-                      <td>
-                        <Mutation
-                          mutation={ADD_USER_TO_TEAM}
-                          refetchQueries={[{ query: USER_LIST, variables: { teamId } }]}
-                        >
-                          {insert_user_team => (
-                            <Button
-                              intent="primary"
-                              icon={IconNames.PLUS}
-                              fill
-                              onClick={() => {
-                                insert_user_team({
-                                  variables: {
-                                    email,
-                                    username,
-                                    teamId,
-                                  },
-                                })
-                                this.setState({
-                                  email: '',
-                                  username: '',
-                                })
-                              }}
-                            />
-                          )}
-                        </Mutation>
-                      </td>
-                    </tr>
-                  </tbody>
-                </table>
-                <div className="container">
-                  <hr className="hr-text" data-content="TRANSFER USER" />
-                  <TeamSelect
-                    eventId={eventId}
-                    teamId={transferId}
-                    handleChange={this.handleTransferSelect}
-                  />
-                  <Mutation
-                    mutation={TRANSFER_USER_TO_TEAM}
-                    refetchQueries={[
-                      { query: USER_LIST, variables: { teamId: transferId } },
-                      { query: USER_LIST, variables: { teamId } },
-                    ]}
-                  >
-                    {update_user_team => (
-                      <Button
-                        intent="primary"
-                        large
-                        style={{ marginTop: 10 }}
-                        onClick={() => {
-                          update_user_team({
-                            variables: {
-                              teamId: transferId,
-                              users: [...checkedItems.keys()],
-                            },
-                          }).then(() =>
-                            this.setState(prevState => ({
-                              checkedItems: new Map(),
-                              transferId: null,
-                            })),
-                          )
-                        }}
-                      >
-                        Confirm
-                      </Button>
-                    )}
-                  </Mutation>
-                </div>
-              </div>
-            )
-          }}
-        </Query>
-      </div>
-    )
-  }
+const initialState = {
+  teamId: null,
+  username: null,
+  email: null,
+  checkedItems: new Map(),
+  transferId: null,
 }
+
+const ManageUsersTab = ({ eventId }) => {
+  const [formState, dispatch] = useReducer(
+    (state, newState) => ({ ...state, ...newState }),
+    initialState,
+  )
+
+  // =================================================
+  // Change Handlers
+  // =================================================
+  const handleSelect = e => {
+    dispatch({ teamId: e.target.value })
+  }
+
+  const handleInputChange = e => {
+    dispatch({ [e.target.name]: e.target.value })
+  }
+
+  const handleCheckBox = e => {
+    const { name, checked } = e.target
+    dispatch({ checkedItems: formState.checkedItems.set(name, checked) })
+  }
+
+  const handleTransferSelect = e => {
+    dispatch({ transferId: e.target.value })
+  }
+
+  const { username, email, teamId, transferId, checkedItems } = formState
+  return (
+    <div>
+      <TeamSelect eventId={eventId} teamId={teamId} handleChange={handleSelect} />
+      <Query query={USER_LIST} variables={{ teamId }}>
+        {({ data, loading, error }) => {
+          if (!data) return null
+          if (loading) return <div>Loading...</div>
+          if (error) return <div>Error...</div>
+
+          if (teamId === null) {
+            return <H4>Select a user from the dropdown</H4>
+          }
+
+          if (!Array.isArray(data.user_team) || !data.user_team.length) {
+            return (
+              <table style={{ width: '100%', marginTop: 10 }}>
+                <thead>
+                  <td>Username</td>
+                  <td>Email</td>
+                  <td />
+                </thead>
+                <tbody>
+                  <tr>
+                    <td>
+                      <InputGroup name="username" value={username} onChange={handleInputChange} />
+                    </td>
+                    <td>
+                      <InputGroup name="email" value={email} onChange={handleInputChange} />
+                    </td>
+                    <td>
+                      <Mutation
+                        mutation={ADD_USER_TO_TEAM}
+                        refetchQueries={[{ query: USER_LIST, variables: { teamId } }]}
+                      >
+                        {insert_user_team => (
+                          <Button
+                            intent="primary"
+                            icon={IconNames.PLUS}
+                            fill
+                            onClick={() => {
+                              insert_user_team({
+                                variables: {
+                                  email,
+                                  username,
+                                  teamId,
+                                },
+                              })
+                              dispatch({
+                                email: '',
+                                username: '',
+                              })
+                            }}
+                          />
+                        )}
+                      </Mutation>
+                    </td>
+                  </tr>
+                </tbody>
+              </table>
+            )
+          }
+
+          return (
+            <div style={{ marginTop: 10 }}>
+              <table style={{ width: '100%' }}>
+                <thead>
+                  <td>Username</td>
+                  <td>Email</td>
+                  <td />
+                </thead>
+                <tbody>
+                  {data.user_team.map(({ user }) => (
+                    <tr key={user.uuid}>
+                      <td>{user.username}</td>
+                      <td>{user.email}</td>
+                      <td>
+                        <Checkbox
+                          name={user.uuid}
+                          checked={checkedItems.get(user.uuid)}
+                          onChange={handleCheckBox}
+                          large
+                          style={{ width: '100%' }}
+                        />
+                      </td>
+                    </tr>
+                  ))}
+                  <tr>
+                    <td>
+                      <InputGroup name="username" value={username} onChange={handleInputChange} />
+                    </td>
+                    <td>
+                      <InputGroup name="email" value={email} onChange={handleInputChange} />
+                    </td>
+                    <td>
+                      <Mutation
+                        mutation={ADD_USER_TO_TEAM}
+                        refetchQueries={[{ query: USER_LIST, variables: { teamId } }]}
+                      >
+                        {insert_user_team => (
+                          <Button
+                            intent="primary"
+                            icon={IconNames.PLUS}
+                            fill
+                            onClick={() => {
+                              insert_user_team({
+                                variables: {
+                                  email,
+                                  username,
+                                  teamId,
+                                },
+                              })
+                              dispatch({
+                                email: '',
+                                username: '',
+                              })
+                            }}
+                          />
+                        )}
+                      </Mutation>
+                    </td>
+                  </tr>
+                </tbody>
+              </table>
+              <div className="container">
+                <hr className="hr-text" data-content="TRANSFER USER" />
+                <TeamSelect
+                  eventId={eventId}
+                  teamId={transferId}
+                  handleChange={handleTransferSelect}
+                />
+                <Mutation
+                  mutation={TRANSFER_USER_TO_TEAM}
+                  refetchQueries={[
+                    { query: USER_LIST, variables: { teamId: transferId } },
+                    { query: USER_LIST, variables: { teamId } },
+                  ]}
+                >
+                  {update_user_team => (
+                    <Button
+                      intent="primary"
+                      large
+                      style={{ marginTop: 10 }}
+                      onClick={() => {
+                        update_user_team({
+                          variables: {
+                            teamId: transferId,
+                            users: [...checkedItems.keys()],
+                          },
+                        }).then(() =>
+                          dispatch({
+                            checkedItems: new Map(),
+                            transferId: null,
+                          }),
+                        )
+                      }}
+                    >
+                      Confirm
+                    </Button>
+                  )}
+                </Mutation>
+              </div>
+            </div>
+          )
+        }}
+      </Query>
+    </div>
+  )
+}
+
+export default ManageUsersTab
