+++ src/pages/Admin/Home/components/CountdownClock/index.js
@@ -1,5 +1,6 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
+
 
 export default class CountdownClock extends React.Component {
   constructor(props) {
